gestion pilotos refactorizado
