+++ frontend/app/Equipos/Pilotos/page.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { GestionPilotos } from "../components/GestionPilotos";
+import { GestionPilotos } from "../../components/Equipos/GestionPilotos";
 
 export default function DefaultTable() {
   return <GestionPilotos></GestionPilotos>;
